+++ ydb/core/persqueue/pqtablet/partition/partition_compactification.cpp
@@ -1,6 +1,6 @@
 #include "partition.h"
 #include "partition_compactification.h"
-#include "partition_log.h"
+#include <ydb/core/persqueue/pqtablet/common/logging.h>
 #include <ydb/core/persqueue/write_meta.h>
 #include "partition_util.h"
