+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -23,6 +23,7 @@ PEERDIR(
     ydb/core/engine/minikql
     ydb/core/persqueue/events
     ydb/core/persqueue/partition_key_range
+    ydb/core/persqueue/pqtablet/common
     ydb/library/logger
 )
